+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'apache2liatrio'
+name             'apache2-liatrio'
 maintainer       'Liatrio'
 maintainer_email 'drew@liatrio.com'
 license          'All rights reserved'
-description      'Installs/Configures apache2liatrio'
+description      'Installs/Configures apache2-liatrio'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'